chore: add firebase libraries and rename app package name
New package name is "com.github.onlynotesswent".
