+++ app/.gitignore
@@ -3,3 +3,4 @@
 /build
 /app/build
 local.properties
+google-services.json
