+++ app/src/androidTest/java/com/github/onlynoteswent/screen/MainScreen.kt
@@ -1,7 +1,7 @@
-package com.android.sample.screen
+package com.github.onlynoteswent.screen
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.android.sample.resources.C
+import com.github.onlynoteswent.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
