+++ app/src/main/java/com/github/onlynoteswent/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.github.onlynoteswent.resources
 
 // Like R, but C
 object C {
