+++ app/src/test/java/com/github/onlynoteswent/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.github.onlynoteswent
 
 import org.junit.Assert.*
 import org.junit.Test
