+++ build.gradle.kts
@@ -2,4 +2,6 @@
 plugins {
     alias(libs.plugins.androidApplication) apply false
     alias(libs.plugins.jetbrainsKotlinAndroid) apply false
+    // alias(libs.plugins.gms) apply false //from bootcamp tutorial, doesn't compile
+    id("com.google.gms.google-services") version "4.4.2" apply false // from firebase creation guide
 }
